Integrating https://github.com/microsoft/mu_basecore/commit/f6b86eec5aeb44b205d44312c99f420a96ada43f and https://github.com/microsoft/mu_basecore/commit/fbb6f18e1c5a342285aafcdee5b53b66bf07989c#diff-df879e5510a384911b647604ef97fa2ddc7a249011be485bd272e1372a13fdf6
