+++ SpamPkg/Library/SmmCpuFeaturesLib/StandaloneMmCpuFeaturesLibStm.inf
@@ -2,12 +2,12 @@
 #  The CPU specific programming for PiSmmCpuDxeSmm module when STM support
 #  is included.
 #
-#  Copyright (c) 2009 - 2018, Intel Corporation. All rights reserved.<BR>
+#  Copyright (c) 2009 - 2023, Intel Corporation. All rights reserved.<BR>
 #  SPDX-License-Identifier: BSD-2-Clause-Patent
 #
 ##
 
-#Override : 00000002 | UefiCpuPkg/Library/SmmCpuFeaturesLib/SmmCpuFeaturesLibStm.inf | 802e9e3d1573bf839ee79069ab73f973 | 2023-11-16T21-18-38 | 6f4fd152cf29b18cc0658146b8f7eca6e92ac7d3
+#Override : 00000002 | UefiCpuPkg/Library/SmmCpuFeaturesLib/SmmCpuFeaturesLibStm.inf | 25e0e8142525fe1e23af3fcb5d9c98d7 | 2024-02-27T23-39-08 | e9e34593db8b2e4506164ef123ccae4f5a9c699f
 
 [Defines]
   INF_VERSION                    = 0x00010005
@@ -46,6 +46,9 @@
   DebugLib
   MmServicesTableLib
   TpmMeasurementLib
+
+[Guids]
+  gSmmBaseHobGuid                ## CONSUMES
 
 [Protocols]
   gEfiMmEndOfDxeProtocolGuid  ## CONSUMES
